update icons for all
